Restructuring of program documents
