+++ src/pdfdeal/__init__.py
@@ -1,3 +1,3 @@
-from .dealpdfs import deal_pdf
+from .FileTools.dealpdfs import deal_pdf
 
 __all__ = ['deal_pdf']
